requirements.txt
accelerate==1.12.0
altair @ file:///C:/miniconda3/conda-bld/altair_1765528043854/work
asttokens @ file:///C:/b/abs_9662ywy9fp/croot/asttokens_1743630464377/work
attrs @ file:///C:/miniconda3/conda-bld/attrs_1762356939015/work
blinker @ file:///C:/miniconda3/conda-bld/blinker_1764332315443/work
Bottleneck @ file:///C:/b/abs_816hr2khp1/croot/bottleneck_1731058648110/work
brotlicffi @ file:///C:/b/abs_f4lixvzmxt/croot/brotlicffi_1736183286104/work
cachetools @ file:///C:/miniconda3/conda-bld/cachetools_1764867553352/work
certifi @ file:///C:/miniconda3/conda-bld/certifi_1763135347537/work/certifi
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click @ file:///C:/miniconda3/conda-bld/click_1764332349001/work
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
comm @ file:///C:/miniconda3/conda-bld/comm_1763119233835/work
contourpy==1.3.2
cycler==0.12.1
decorator @ file:///C:/miniconda3/conda-bld/decorator_1757341256088/work
exceptiongroup @ file:///C:/miniconda3/conda-bld/exceptiongroup_1761560437674/work
executing @ file:///C:/miniconda3/conda-bld/executing_1757061252959/work
filelock @ file:///C:/b/abs_758187j281/croot/filelock_1744281404850/work
fonttools==4.60.1
fsspec==2025.10.0
gitdb @ file:///C:/b/abs_1bkmriimss/croot/gitdb_1753362133190/work
GitPython @ file:///C:/miniconda3/conda-bld/gitpython_1761291330593/work
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
huggingface-hub==0.36.0
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
ipython @ file:///C:/b/abs_8eyhzleyrk/croot/ipython_1734548134403/work
ipywidgets @ file:///C:/miniconda3/conda-bld/ipywidgets_1761560181476/work
jedi @ file:///C:/b/abs_3a2kbnlclc/croot/jedi_1733987412687/work
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
jsonschema @ file:///C:/miniconda3/conda-bld/jsonschema_1762459331642/work
jsonschema-specifications @ file:///C:/miniconda3/conda-bld/jsonschema-specifications_1762788648330/work
jupyterlab_widgets @ file:///C:/b/abs_6566by8dag/croot/jupyterlab_widgets_1752066574111/work
kiwisolver==1.4.9
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
matplotlib==3.10.7
matplotlib-inline @ file:///C:/miniconda3/conda-bld/matplotlib-inline_1762779265381/work
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
narwhals @ file:///C:/miniconda3/conda-bld/narwhals_1760003512045/work
networkx @ file:///C:/b/abs_b054htfn9t/croot/networkx_1737043671910/work
numexpr @ file:///C:/b/abs_4e0h6i5ip2/croot/numexpr_1752521729801/work
numpy @ file:///C:/b/abs_0123vcxhf8/croot/numpy_and_numpy_base_1725470331966/work/dist/numpy-2.0.1-cp310-cp310-win_amd64.whl#sha256=e824ec3b279f4c5207736c83ac2f2dc79b489e08501a4ea9e3fc178c45db289b
opencv-python==4.12.0.88
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049101700/work
pandas @ file:///C:/miniconda3/conda-bld/pandas_1759822263292/work/dist/pandas-2.3.3-cp310-cp310-win_amd64.whl#sha256=1334d37cdb28f0f10d322297a276c0b56e115ad32f5bf3f0e575560f61a5ea74
parso @ file:///C:/miniconda3/conda-bld/parso_1762781925562/work
pillow @ file:///C:/b/abs_b50vowcrzo/croot/pillow_1738010273782/work
polars==1.35.1
polars-runtime-32==1.35.1
prompt_toolkit @ file:///C:/miniconda3/conda-bld/prompt-toolkit_1761745021230/work
protobuf @ file:///C:/b/abs_b5v84rwh56/croot/protobuf_1742419721750/work/bazel-bin/python/dist/protobuf-5.29.3-cp310-abi3-win_amd64.whl#sha256=dd7a6e96ccf2f19dc63e35889d44865cfe6df0449c799d650a73fb28d3073c4c
psutil==7.1.2
pure_eval @ file:///C:/miniconda3/conda-bld/pure_eval_1757067068391/work
pyarrow @ file:///C:/miniconda3/conda-bld/pyarrow_1764939502439/work/python
pycparser @ file:///C:/miniconda3/conda-bld/pycparser_1757496059965/work
Pygments @ file:///C:/miniconda3/conda-bld/pygments_1762431425987/work
pyparsing==3.2.5
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
python-dateutil @ file:///C:/b/abs_3au_koqnbs/croot/python-dateutil_1716495777160/work
pytorch-fid==0.3.0
pytz @ file:///C:/b/abs_f8wdzeix0n/croot/pytz_1752135878094/work
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
referencing @ file:///C:/miniconda3/conda-bld/referencing_1762536910091/work
regex==2025.11.3
requests @ file:///C:/miniconda3/conda-bld/requests_1762359612241/work
rpds-py @ file:///C:/miniconda3/conda-bld/rpds-py_1762366614812/work
safetensors==0.7.0
scipy==1.15.3
six @ file:///C:/b/abs_149wuyuo1o/croot/six_1744271521515/work
smmap @ file:///C:/b/abs_85z04amxrn/croot/smmap_1752748960667/work
stack_data @ file:///C:/miniconda3/conda-bld/stack_data_1757067051465/work
streamlit @ file:///C:/Users/task_176605006251550/croot/streamlit-split_1766050507829/work/streamlit
sympy==1.13.1
tenacity @ file:///C:/miniconda3/conda-bld/tenacity_1760375743144/work
tokenizers==0.19.1
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado @ file:///C:/Users/task_176599267555405/croot/tornado_1765992779662/work
tqdm==4.67.1
traitlets @ file:///C:/b/abs_bfsnoxl4pq/croot/traitlets_1718227069245/work
transformers==4.41.1
typing_extensions @ file:///C:/b/abs_ecq8gc0vbm/croot/typing_extensions_1756281142218/work
tzdata @ file:///croot/python-tzdata_1746123641790/work
ultralytics==8.3.223
ultralytics-thop==2.0.18
urllib3 @ file:///C:/b/abs_393g1ayee_/croot/urllib3_1750775481226/work
watchdog @ file:///C:/miniconda3/conda-bld/watchdog_1759775793959/work
wcwidth @ file:///C:/b/abs_0c8p4c33bp/croot/wcwidth_1750352902378/work
widgetsnbextension @ file:///C:/miniconda3/conda-bld/widgetsnbextension_1761556853993/work
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
